trying to clean up the main __init__
